requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
gunicorn==22.0.0
pydantic==2.8.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sentence-transformers==3.0.1
torch==2.3.1
faiss-cpu==1.8.0.post1
